requirements.txt
Flask
ultralytics
opencv-python
paho-mqtt
python-dotenv
requests
dotenv